Remove white space before stylesheet header content
This is required so the ‘Tags’ line does not fail the max-line-length
rule in the stylelint-config-wordpress we extend from in .stylelintrc.

If stylelint is run when there's space before the ‘Tags’ line, we see
‘Expected line length to be no more than 80 characters’ for that line.

The WordPress stylelint rules contain a pattern to ignore line length
for the stylesheet header, but only if the ‘Tags’ line is not preceded
by white space: https://tinyurl.com/yb3xb5n6.