Drop deprecated CHILD_THEME_NAME constant and use CHILD_THEME_HANDLE
CHILD_THEME_NAME is no longer required as of Genesis 2.9.0. It was
previously used to set the footer credit text and as the handle for
the enqueued stylesheet. In Genesis 2.9.0 the name is now pulled from
the Theme Name header for the site footer, and a sanitized version
of the same value is used as the stylesheet handle.

See https://github.com/studiopress/genesis/pull/2166.

The CHILD_THEME_HANDLE constant is used instead for enqueueing scripts
and styles. Where a human-readable name is needed (admin notices), it is
pulled directly from the stylesheet Theme Name line.

The goal is to reduce the number of places where people need to set
their theme name and version number.

The sanitized Name is used for the CHILD_THEME_HANDLE instead of the
Text Domain because we have often seen modified Genesis Sample themes
where the name is changed but the text domain is not. We could switch to
the Text Domain once this is implemented:
https://github.com/studiopress/genesis-sample/issues/142.

It's also likely that we'll switch away from constants and use helper
functions should those become available in Genesis. This would offer
better cache-busting during development, where assets can change often
but version strings do not.
https://github.com/studiopress/genesis/pull/2199#issuecomment-470190883.