Updates/wordpress 53 (#318)
* Account for gallery class change

WordPress 5.3 moves the `wp-block-gallery` class to the figure rather than the ul and adds `blocks-gallery-grid` to the ul. This updates to account for the class change.

* Add center alignement to figcaption

WP 5.3 has removed `text-align: center` from block styles for figcaption. This adds it back to restore the original style and prevent full width captions from being flush with the browser edge.

* Add styling for gallery captions

* Remove unnecessary button CSS

WordPress 5.3 removes the squared button style in favor of a border-radius setting. The default border-radius setting is blank in the editor. Thus, the default of 5px is left set in the current CSS. We could possibly move this to inline styles and set it in the config.

* Update caption styles for editor

Removes duplicate figcaption style as well.

* Remove custom wp-block-column CSS

WP 5.3 removes the `has-x-columns` classes so they are no longer relevant. Also, defer to default margin on columns.

* Remove border from is-style-stripes table style

* Allow fixed table block setting to work

* Adjust table padding

* Add alignment styles

WP 5.3 adds `has-text-align-center`, `has-text-align-left`, & `has-text-align-right` classes. In some cases, these are styled by WP, but in others, the styles are missing.

* Match editor and front end table styles

* Update homepage content import

- Update for alignment classes
- Change from WP Columns block to AB Advanced Columns

* Match width of inner blocks to front end

* Allow 5.3 hr color option

* Updates/53 width (#317)

* Updates for nested blocks

* AB Spacer background transparency

* Update editor separator styles

* Alternate separator method

* Backward compatibility for gallery padding

* Consolidate table styles

* Editor width adjustment

* Revert "Editor width adjustment"

This reverts commit c81ec79c328ba8c4d0b7ea90bdbedd039fd84031.

* Combine rules

* Reorder CSS

* Update category/featured post block

Also allows child option to work in the category block.

* Minor editor adjustment

* Adjust full-width cover block width

* Scope CSS to .entry-content

* Consolidate and better match captions

* Adjust column media query

* Update version and changelog

* Changelog addition

* Updates/genesis 32 (#319)

* Set post meta on theme activation

* Allow footer widgets on landing page template

* Hide footer widgets on imported landing page